fix(motion): T6-port-b949cebf — PR #197 CI fixes (clang-format, clang-tidy, doc-substance, AGENTS.md)
Clears the 4 CI failures that arrived after the upstream port commit:

1. Pre-Commit (clang-format) — re-flow motion.c hunks under the
   100-column / single-space K&R style. No semantic delta.
2. Clang-Tidy — picture_copy.h was missing the libvmaf/picture.h
   include and tripped clang-diagnostic-error on VmafPicture; add
   the include. Drop redundant trailing returns in
   picture_copy_hbd() / picture_copy(). Cast outer index to
   ptrdiff_t in motion.c / float_motion.c / float_psnr.c inner
   loops to silence bugprone-implicit-widening-of-multiplication.
   Mark unused bpc parameter in float_motion init() with (void).
   Split unsigned h_u, h_v / double score_u, score_v / double
   score2_u, score2_v multi-decls (readability-isolate-decl).
   NOLINTNEXTLINE(readability-function-size) on the 3 large
   float_motion.c functions (init/extract/close) the upstream
   port grows inline + the pre-existing float_adm.c extract()
   touched only by signature update; cite ADR-0141 / T7-5 sweep.
3. Doc-Substance Gate — add motion-port note + new option matrix
   row to docs/metrics/features.md, and "Known gaps" entries for
   the CPU-only motion_add_uv path to docs/backends/cuda/overview.md
   and docs/backends/sycl/overview.md (the surface gate flagged
   touched cuda/ + sycl/ feature kernels without matching docs).
4. Deep-Dive Deliverables Checklist — add a libvmaf/AGENTS.md
   rebase-sensitive-invariant entry pinning the float_motion
   extra-options surface, the SIMD-fast-path wrapper, and the
   picture_copy() trailing-channel signature change. PR body
   refresh in the same push lands the matching ticked rows.

No code path is exercised differently — defaults stay bit-identical
to pre-port baseline (`compute_motion_simd` Y-plane fast path); the
NOLINT additions cite the load-bearing invariant per CLAUDE.md §12 r12.

+++ libvmaf/src/feature/motion.c
@@ -34,10 +34,10 @@
 #include "motion_tools.h"
 
 #define convolution_f32_c convolution_f32_c_s
-#define FILTER_3           FILTER_3_s
-#define FILTER_5           FILTER_5_s
-#define FILTER_5_NO_OP     FILTER_5_NO_OP_s
-#define offset_image       offset_image_s
+#define FILTER_3 FILTER_3_s
+#define FILTER_5 FILTER_5_s
+#define FILTER_5_NO_OP FILTER_5_NO_OP_s
+#define offset_image offset_image_s
 
 #define MIN(x, y) (((x) < (y)) ? (x) : (y))
 
@@ -63,8 +63,7 @@ static float motion_bilinear_interp(const float *src, int width, int height, int
 
     return ((1.0f - dy) * (1.0f - dx) * src[y1 * src_stride + x1] +
             (1.0f - dy) * dx * src[y1 * src_stride + x2] +
-            dy * (1.0f - dx) * src[y2 * src_stride + x1] +
-            dy * dx * src[y2 * src_stride + x2]);
+            dy * (1.0f - dx) * src[y2 * src_stride + x1] + dy * dx * src[y2 * src_stride + x2]);
 }
 
 static void motion_scale_bilinear(const float *src, float *dst, int src_w, int src_h,
@@ -82,8 +81,7 @@ static void motion_scale_bilinear(const float *src, float *dst, int src_w, int s
         float yy = (y + 0.5f) * ratio_y - 0.5f;
         for (int x = 0; x < dst_w; x++) {
             float xx = (x + 0.5f) * ratio_x - 0.5f;
-            dst[y * dst_stride + x] =
-                motion_bilinear_interp(src, src_w, src_h, src_stride, xx, yy);
+            dst[y * dst_stride + x] = motion_bilinear_interp(src, src_w, src_h, src_stride, xx, yy);
         }
     }
 }
@@ -91,8 +89,8 @@ static void motion_scale_bilinear(const float *src, float *dst, int src_w, int s
 /**
  * Note: img1_stride and img2_stride are in terms of (sizeof(float) bytes)
  */
-float vmaf_image_sad_c(const float *img1, const float *img2, int width, int height,
-                       int img1_stride, int img2_stride, int motion_add_scale1)
+float vmaf_image_sad_c(const float *img1, const float *img2, int width, int height, int img1_stride,
+                       int img2_stride, int motion_add_scale1)
 {
     float motion_scale0 = 0.0;
     float accum = (float)0.0;
@@ -126,8 +124,8 @@ float vmaf_image_sad_c(const float *img1, const float *img2, int width, int heig
         for (int i = 0; i < scaled_height; ++i) {
             float accum_line = (float)0.0;
             for (int j = 0; j < scaled_width; ++j) {
-                float img1px = img1_scaled[i * scaled_float_stride / sizeof(float) + j];
-                float img2px = img2_scaled[i * scaled_float_stride / sizeof(float) + j];
+                float img1px = img1_scaled[(size_t)i * scaled_float_stride / sizeof(float) + j];
+                float img2px = img2_scaled[(size_t)i * scaled_float_stride / sizeof(float) + j];
                 accum_line += fabsf(img1px - img2px);
             }
             accum_scale1 += accum_line;
@@ -147,8 +145,8 @@ float vmaf_image_sad_c(const float *img1, const float *img2, int width, int heig
 /**
  * Note: ref_stride and dis_stride are in terms of bytes
  */
-int compute_motion(const float *ref, const float *dis, int w, int h, int ref_stride,
-                   int dis_stride, double *score, int motion_decimate)
+int compute_motion(const float *ref, const float *dis, int w, int h, int ref_stride, int dis_stride,
+                   double *score, int motion_decimate)
 {
     if (ref_stride % sizeof(float) != 0) {
         printf("error: ref_stride %% sizeof(float) != 0, ref_stride = %d, sizeof(float) = %zu.\n",
